+++ src/server/core/ml/__init__.py
@@ -1,50 +1,6 @@
-import nltk
-from nltk.corpus import sentiwordnet as swn
 from src.server.core import SingletonException, AnalyzerStrategy
 from src.server.preprocessor import PreProcessor
 from src.server.core.ml.bayes import BayesBank
-
-
-class SentiWordNet:
-    __instance = None
-
-    @staticmethod
-    def get_instance():
-        if SentiWordNet.__instance is None:
-            SentiWordNet()
-        return SentiWordNet.__instance
-
-    def __init__(self):
-        if SentiWordNet.__instance is not None:
-            raise SingletonException
-        else:
-            try:
-                nltk.data.find('corpora/wordnet.zip/wordnet/')
-                nltk.data.find('corpora/sentiwordnet.zip/sentiwordnet/')
-            except:
-                nltk.download('wordnet')
-                nltk.download('sentiwordnet')
-                print(nltk.data.find('corpora/wordnet.zip/wordnet/'))
-                print(nltk.data.find('corpora/sentiwordnet.zip/sentiwordnet/'))
-            SentiWordNet.__instance = self
-
-    def senti_query(self, word):
-        results = list(swn.senti_synsets(word))
-        pos = 0
-        neg = 0
-        count = 0
-
-        for result in results:
-            if result.synset.name().split(".")[0] == word:
-                pos += result.pos_score()
-                neg += result.neg_score()
-                count += 1
-
-        if count > 0:
-            return neg / count - pos / count
-        else:
-            return 0
-
 
 class MachineLearning(AnalyzerStrategy):
     def __init__(self):
